Move more functionality into task
