+++ packages/op-datasets/src/op_datasets/etl/ingestion/sources.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
+
 from op_coreutils.logger import structlog
-
-from op_datasets.pipeline.ozone import BlockBatch
-from op_datasets.schemas import CoreDataset
-
 from overrides import EnforceOverrides, override
 
+from op_datasets.etl.ingestion.batches import BlockBatch
+from op_datasets.schemas import CoreDataset
+
+from . import fromgoldsky, fromlocal
 
 log = structlog.get_logger()
 
@@ -37,8 +38,6 @@ class GoldskySource(CoreDatasetSource):
         datasets: dict[str, CoreDataset],
         block_batch: BlockBatch,
     ):
-        from op_datasets.coretables import fromgoldsky
-
         return fromgoldsky.read_core_tables(datasets, block_batch)
 
 
@@ -52,6 +51,4 @@ class LocalFileSource(CoreDatasetSource):
         datasets: dict[str, CoreDataset],
         block_batch: BlockBatch,
     ):
-        from op_datasets.coretables import fromlocal
-
         return fromlocal.read_core_tables(self.basepath, datasets, block_batch)
