+++ src/op_analytics/cli/subcommands/chains/app.py
@@ -11,8 +11,8 @@ from op_datasets.chains.chain_metadata import (
     to_pandas,
 )
 from op_datasets.etl.ingestion import ingest
+from op_datasets.etl.ingestion.batches import split_block_range
 from op_datasets.pipeline.blockrange import BlockRange
-from op_datasets.pipeline.ozone import split_block_range
 from op_datasets.schemas import ONCHAIN_CURRENT_VERSION
 from rich import print
 from typing_extensions import Annotated
